+++ sdk/Managed/src/Microsoft.WindowsAzure.MobileServices.WindowsPhone8/Platform/ApplicationStorage.cs
@@ -3,7 +3,6 @@
 // ----------------------------------------------------------------------------
 
 using System;
-using System.Globalization;
 using System.IO.IsolatedStorage;
 
 namespace Microsoft.WindowsAzure.MobileServices
@@ -18,7 +17,16 @@ namespace Microsoft.WindowsAzure.MobileServices
         /// <summary>
         /// A singleton instance of the <see cref="ApplicationStorage"/>.
         /// </summary>
-        private static IApplicationStorage instance = new ApplicationStorage();
+        private static readonly IApplicationStorage instance = new ApplicationStorage();        
+
+        private ApplicationStorage() : this(string.Empty)
+        {
+        }
+
+        internal ApplicationStorage(string name)
+        {
+            this.StoragePrefix = name;
+        }
 
         /// <summary>
         /// A singleton instance of the <see cref="ApplicationStorage"/>.
@@ -30,6 +38,8 @@ namespace Microsoft.WindowsAzure.MobileServices
                 return instance;
             }
         }
+
+        private string StoragePrefix { get; set; }
 
         /// <summary>
         /// Tries to read a setting's value from 
@@ -54,7 +64,7 @@ namespace Microsoft.WindowsAzure.MobileServices
                 throw new ArgumentException(message);
             }
 
-            return IsolatedStorageSettings.ApplicationSettings.TryGetValue<object>(name, out value);
+            return IsolatedStorageSettings.ApplicationSettings.TryGetValue(string.Concat(this.StoragePrefix, name), out value);
         }
 
         /// <summary>
@@ -75,7 +85,12 @@ namespace Microsoft.WindowsAzure.MobileServices
                 throw new ArgumentException(message);
             }
 
-            IsolatedStorageSettings.ApplicationSettings[name] = value;
+            IsolatedStorageSettings.ApplicationSettings[string.Concat(this.StoragePrefix, name)] = value;
+        }
+
+        void IApplicationStorage.Save()
+        {
+            IsolatedStorageSettings.ApplicationSettings.Save();
         }
     }
 }