+++ sdk/Managed/src/Microsoft.WindowsAzure.MobileServices.WindowsStore/Platform/PlatformInformation.cs
@@ -15,7 +15,7 @@ namespace Microsoft.WindowsAzure.MobileServices
         /// <summary>
         /// A singleton instance of the <see cref="IPlatformInformation"/>.
         /// </summary>
-        private static IPlatformInformation instance = new PlatformInformation();
+        private static readonly IPlatformInformation instance = new PlatformInformation();
 
         /// <summary>
         /// A singleton instance of the <see cref="IPlatformInformation"/>.