+++ sdk/Managed/test/Microsoft.WindowsAzure.MobileServices.Test/FunctionalTests/QueryingTest.cs
@@ -67,6 +67,7 @@ namespace Microsoft.WindowsAzure.MobileServices.Test
                 .OrderByDescending(p => p.Title));
         }
 
+        [Tag("notXamarin_iOS")]
         [AsyncTestMethod]
         public async Task LiveProjection()
         {