+++ sdk/Managed/test/Microsoft.WindowsAzure.MobileServices.Test/FunctionalTests/Table/MobileServiceTableTest.cs
@@ -63,13 +63,22 @@ namespace Microsoft.WindowsAzure.MobileServices.Test
         {
             // Make sure the table is empty
             IMobileServiceTable<T> table = GetClient().GetTable<T>();
-            table.MobileServiceClient.SerializerSettings.DateParseHandling = DateParseHandling.None;
-            IEnumerable<T> results = await table.ReadAsync();
-            T[] items = results.ToArray();
+            table.MobileServiceClient.SerializerSettings.DateParseHandling = DateParseHandling.None;            
 
-            foreach (T item in items)
+            while (true)
             {
-                await table.DeleteAsync(item);
+                IEnumerable<T> results = await table.Take(1000).ToListAsync();
+                T[] items = results.ToArray();
+
+                if (!items.Any())
+                {
+                    break;
+                }
+
+                foreach (T item in items)
+                {
+                    await table.DeleteAsync(item);
+                }
             }
         }